LayoutTests/platform/ios/imported/w3c/web-platform-tests/css/css-grid/masonry/tentative/track-sizing/auto-repeat/row-auto-repeat-009-expected.txt


FAIL .masonry 1 assert_equals:
<div class="masonry" data-expected-width="100" data-expected-height="500">
      <div class="item" data-expected-width="100" data-expected-height="100">Cell 1</div>
      <div class="item" data-expected-width="100" data-expected-height="100">Cell 2</div>
      <div class="item" data-expected-width="100" data-expected-height="100">Cell 3</div>
      <div class="item" data-expected-width="100" data-expected-height="100">Cell 4</div>
      <div class="item" data-expected-width="100" data-expected-height="100">Cell 5</div>
    </div>
height expected 100 but got 20
Cell 1
Cell 2
Cell 3
Cell 4
Cell 5
